Cast backlog explicitely from unsigned int to int.

+++ src/linux/cio_linux_server_socket.c
@@ -59,7 +59,7 @@ static enum cio_error socket_init(void *context, unsigned int backlog)
 		return err;
 	}
 
-	ss->backlog = backlog;
+	ss->backlog = (int)backlog;
 	ss->ev.fd = listen_fd;
 
 	return cio_success;
